[feat] CoreDataStack 저장 기능 구현 CoreDataStack 저장 기능을 Singleton 구조로 구현
