+++ Project17-C-Map/InteractiveClusteringMap/Core Data/DataManagable.swift
@@ -1,0 +1,16 @@
+//
+//  DataManagable.swift
+//  InteractiveClusteringMap
+//
+//  Created by Oh Donggeon on 2020/11/21.
+//
+
+import Foundation
+
+protocol DataManagable {
+    
+    func fetch()
+    func save(successHandler: @escaping () -> Void, failureHandler: ((NSError) -> Void)?)
+    func setValue()
+    
+}
